requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.0
Flask==1.1.1
Flask-HTTPAuth==3.3.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
plivo==4.3.2
python-dotenv==0.10.3
redis==3.3.11
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
nexmo==2.4.0
logdna==1.4.2
pprint==0.1
uuid==1.30
phonenumbers==8.10.22